% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs-package.R
\docType{package}
\name{fmrs-package}
\alias{fmrs}
\alias{fmrs-package}
\title{fmrs: Variable Selection in Finite Mixture of AFT Regression and FMR}
\description{
The fmrs package provides several useful functions. The
    function \code{\link{fmrsmle}} provides Maximum Likelihood estimation
    for FMRs models. The function \code{\link{fmrstunsel}} provides
    component-wise tuning parameters. The function
    \code{\link{fmrsvarsel}} provides variable selection for FMRs models.
}
\section{\code{fmrs} functions}{

\code{\link{fmrsmle}},
\code{\link{fmrstunsel}},
\code{\link{fmrsvarsel}},
\code{\link{fmrsgendata}}.
}

