% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.lamb.object.R
\docType{class}
\name{fmrs.lambda-class}
\alias{fmrs.lambda-class}
\alias{frms.lambda}
\title{An S4 class to represent estimated optimal lambdas}
\description{
An S4 class to represent estimated optimal lambdas resulted from runnig \code{\link{fmrs.tunsel}}.
}
\section{Slots}{

\describe{
\item{\code{lamPen}}{A length-\code{nComp} numeric vector}

\item{\code{disFamily}}{A length-one character vector}

\item{\code{penFamily}}{A length-one character vector}

\item{\code{lamRidge}}{A length-one numeric vector}

\item{\code{method}}{A length-one character vector}

\item{\code{data}}{A list including \code{y}, \code{x} and \code{delta}}
}}

