% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{mixProp}
\alias{mixProp}
\alias{mixProp,fmrsfit-method}
\title{Method mixProp}
\usage{
mixProp(object, ...)

\S4method{mixProp}{fmrsfit}(object, ...)
}
\arguments{
\item{object}{An \code{\link{fmrsfit-class}} object}

\item{...}{other possible arguments}
}
\value{
A numeric array
}
\description{
mixProp method
}
\examples{
set.seed(1980)
nComp = 2
nCov = 10
n = 500
deviance = c(1, 1)
mixProp = c(0.4, 0.6)
rho = 0.5
coeff1 = c( 2,  2, -1, -2, 1, 2, 0, 0,  0, 0,  0)
coeff2 = c(-1, -1,  1,  2, 0, 0, 0, 0, -1, 2, -2)
umax = 40

dat <- fmrsgendata(n = n, nComp = nComp, nCov = nCov,
                     coeff = c(coeff1, coeff2), deviance = deviance,
                     mixProp = mixProp, rho = rho, umax = umax,
                     disFamily = "lnorm")

res.mle <- fmrsmle(y = dat$y, x = dat$x, delta = dat$delta,
                   nComp = nComp, disFamily = "lnorm",
                   initCoeff = rnorm(nComp*nCov+nComp),
                   initDeviance = rep(1, nComp),
                   initmixProp = rep(1/nComp, nComp))
mixProp(res.mle)
}

