% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,fmrsfit-method}
\alias{show,show-method}
\title{show method}
\usage{
show(object)

\S4method{show}{fmrsfit}(object)
}
\arguments{
\item{object}{An \code{\link{fmrsfit-class}}}
}
\value{
Information about the fitted FMRs model
}
\description{
Provides information about the fitted FMRs model
}
\examples{
set.seed(1980)
nComp = 2
nCov = 10
nObs = 500
deviance = c(1, 1)
mixProp = c(0.4, 0.6)
rho = 0.5
coeff1 = c( 2,  2, -1, -2, 1, 2, 0, 0,  0, 0,  0)
coeff2 = c(-1, -1,  1,  2, 0, 0, 0, 0, -1, 2, -2)
umax = 40

dat <- fmrs.gendata(nObs = nObs, nComp = nComp, nCov = nCov,
                     coeff = c(coeff1, coeff2), deviance = deviance,
                     mixProp = mixProp, rho = rho, umax = umax,
                     disFamily = "lnorm")

res.mle <- fmrs.mle(y = dat$y, x = dat$x, delta = dat$delta,
                   nComp = nComp, disFamily = "lnorm",
                   initCoeff = rnorm(nComp*nCov+nComp),
                   initDeviance = rep(1, nComp),
                   initmixProp = rep(1/nComp, nComp))
show(res.mle)
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}

