% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{tunepar-class}
\alias{tunepar}
\alias{tunepar-class}
\title{An S4 class to represent estimated optimal lambdas}
\description{
An S4 class to represent estimated optimal lambdas resulted
    from runnig \code{\link{fmrstunsel}}.
}
\section{Slots}{

\describe{
\item{\code{lamPen}}{A length-\code{nComp} numeric vector}

\item{\code{disFamily}}{A length-one character vector}

\item{\code{penFamily}}{A length-one character vector}

\item{\code{lamRidge}}{A length-one numeric vector}

\item{\code{method}}{A length-one character vector}

\item{\code{data}}{A list including \code{y}, \code{x} and \code{delta}}
}}

