% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{coefficients}
\alias{coefficients}
\alias{coefficients,coefficients-method}
\alias{coefficients,fmrsfit-method}
\title{coefficients method}
\usage{
coefficients(object, ...)

\S4method{coefficients}{fmrsfit}(object, ...)
}
\arguments{
\item{object}{An \code{\link{fmrsfit-class}}}

\item{...}{other possible arguments}
}
\value{
A numeric array of dimension \code{(nCov+1)} by \code{nComp}
}
\description{
Provides the estimated regression coefficients from an
}
\examples{
set.seed(1980)
nComp = 2
nCov = 10
nObs = 500
deviance = c(1, 1)
mixProp = c(0.4, 0.6)
rho = 0.5
coeff1 = c( 2,  2, -1, -2, 1, 2, 0, 0,  0, 0,  0)
coeff2 = c(-1, -1,  1,  2, 0, 0, 0, 0, -1, 2, -2)
umax = 40

dat <- fmrs.gendata(nObs = nObs, nComp = nComp, nCov = nCov,
                     coeff = c(coeff1, coeff2), deviance = deviance,
                     mixProp = mixProp, rho = rho, umax = umax,
                     disFamily = "lnorm")

res.mle <- fmrs.mle(y = dat$y, x = dat$x, delta = dat$delta,
                   nComp = nComp, disFamily = "lnorm",
                   initCoeff = rnorm(nComp*nCov+nComp),
                   initDeviance = rep(1, nComp),
                   initmixProp = rep(1/nComp, nComp))
coefficients(res.mle)
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
\code{\link{fmrsfit-class}}
}

