% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs-package.R
\docType{package}
\name{fmrs-package}
\alias{fmrs}
\alias{fmrs-package}
\title{Variable Selection in Finite Mixture of AFT Regression and
    FMR}
\description{
Provides parameter estimation and variable selection in
    FMRs models. The
    \code{\link{fmrs.mle}} method provides Maximum Likelihood Estimation
    for FMRs models. The \code{\link{fmrs.tunsel}} method provides
    component-wise tuning parameters. The \code{\link{fmrs.varsel}}
    method provides variable selection for FMRs models.
}
\section{\code{fmrs} methods}{

\code{\link{fmrs.mle}},
\code{\link{fmrs.tunsel}},
\code{\link{fmrs.varsel}},
\code{\link{fmrs.gendata}}.
}

\section{\code{fmrs} objects}{

\code{\link{fmrsfit-class}},
\code{\link{fmrstunpar-class}}
}

