% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMR_Surv_MLE.R
\name{fmrs.mle}
\alias{fmrs.mle}
\title{Maximum Likelihood Estimation in Finite Mixture of Accelerated Failure Time Regression Models
and Finite Mixture of Regression Models.}
\usage{
fmrs.mle(y, x, delta, nComp, disFamily = "lnorm", initCoeff, initSigma,
  initPi, lambRidge = 0, nIterEM = 2000, nIterNR = 2, conveps = 1e-08,
  convepsEM = 1e-08, convepsNR = 1e-08, porNR = 2)
}
\arguments{
\item{y}{Time-to-event observations}

\item{x}{Design matrix (covariates)}

\item{delta}{Censoring indicator vector}

\item{nComp}{Orde (Number of components) of mixture model}

\item{disFamily}{Specify sub-distributions family. The options are \code{"norm"} for FMR models,
\code{"lnorm"} for mixture of AFT regression models with Log-Normal sub-distributions,
\code{"weibull"} for mixture of AFT regression models with Weibull sub-distributions,}

\item{initCoeff}{Vector of initial values for coefficients including intercepts.}

\item{initSigma}{Vector of initial values for standard deviations}

\item{initPi}{Vector of initial values for proportion of components}

\item{lambRidge}{Lambda for ridge penalty}

\item{nIterEM}{Number of iteration for EM algorithm}

\item{nIterNR}{Number of iteration for Newton-Raphson algorithm}

\item{conveps}{A positive number for avoiding NaN in computing divisions.}

\item{convepsEM}{Treshold for convergence of EM algorithm}

\item{convepsNR}{Treshold for convergence of Newton-Raphson algorithm}

\item{porNR}{Used in pow(0.5, porNR) for tuning the increment in Newton-Raphson algorithm.}
}
\value{
A fitted FMRs model object of class \code{\link{fmrs.fit}}
}
\description{
It provides parameter estimation for Finite Mixture of Accelerated Failure Time Regression Models
and Finite Mixture of Regression Models.
It also provide Ridge Regression.
}
\examples{
\dontrun{ Ovarian Cancer analysis


}
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
\references{
Shokoohi, F., Khalili, A., Asgharian, M. and Lin, S. (2016) Variable Selection in Mixture of Survival Models
}
\seealso{
Other lnorm..norm..weibull: \code{\link{fmrs.gen.data}},
  \code{\link{fmrs.tunsel}}, \code{\link{fmrs.varsel}}
}
\keyword{AFT,}
\keyword{Algorithm,}
\keyword{Censored}
\keyword{Data,}
\keyword{EM}
\keyword{FMR,}
\keyword{Regression}
\keyword{Ridge}

