% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\docType{package}
\name{fmrs}
\alias{fmrs}
\alias{fmrs-package}
\title{fmrs: Variable Selection in Finite Mixture of Accelerated Failure Time Regression Models and Finite Mixture of Regression Models}
\description{
The fmrs package provides several useful functions. The function \code{fmrs.mle()} provides Maximum Likelihood estimation for FMRs models.
The function \code{fmrs.tunsel()} provides component-wise tuning parameters. The function \code{fmrs.varsel()} provides variable selection for FMRs models.
}
\section{fmrs functions}{

\code{fmrs.mle()},
\code{fmrs.tunsel()},
\code{fmrs.varsel()},
\code{fmrs.gen.data()}.
}

