% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMR_Surv_CwTuneSel.R
\name{fmrs.tunsel}
\alias{fmrs.tunsel}
\title{Component-wise Tuning Parameter Selection in Finite Mixture of Accelerated Failure Time Regression Models
and Finite Mixture of Regression Models}
\usage{
fmrs.tunsel(y, x, delta, nComp, disFamily = "lnorm", initCoeff, initSigma,
  initPi, penFamily = "lasso", lambRidge = 0, nIterEM = 2000,
  nIterNR = 2, conveps = 1e-08, convepsEM = 1e-08, convepsNR = 1e-08,
  porNR = 2, gamMixPor = 1)
}
\arguments{
\item{y}{Responses (observations)}

\item{x}{Design matrix (covariates)}

\item{delta}{Censoring indicator vector}

\item{nComp}{Order (Number of components) of mixture model}

\item{disFamily}{Specify sub-distributions family. The options are \code{"norm"} for FMR models,
\code{"lnorm"} for mixture of AFT regression models with Log-Normal sub-distributions,
\code{"weibull"} for mixture of AFT regression models with Weibull sub-distributions,}

\item{initCoeff}{Vector of initial values for regression coefficients including intercepts}

\item{initSigma}{Vector of initial values for standard deviations}

\item{initPi}{Vector of initial values for proportion of components}

\item{penFamily}{Name of the penalty that is used in variable selection method.
The available options are  \code{"lasso"}, \code{"adplasso"}, \code{"mcp"}, \code{"scad"}, \code{"sica"} and \code{"hard"}.}

\item{lambRidge}{A positive value for Lambda in Ridge regression or Elastic Net}

\item{nIterEM}{Maximum number of iterations for EM algorithm}

\item{nIterNR}{Maximum number of iterations for Newton-Raphson algorithm}

\item{conveps}{A positive value for avoiding NaN in computing divisions}

\item{convepsEM}{A positive value for treshold of convergence in EM algorithm}

\item{convepsNR}{A positive value for treshold of convergence in Newton-Raphson algorithm}

\item{porNR}{Used in pow(0.5, porNR) for tuning the increment in Newton-Raphson algorithm.}

\item{gamMixPor}{Proportion of mixing parameters in the penalty. The value must be in the interval [0,1]. If \code{gamMixPor = 0}, the penalty structure is no longer mixture.}
}
\value{
An \code{\link{fmrs.lambda}} object which includes component-wise tuning parameter estimates to be used in variable selection procedure.
}
\description{
It provides component-wise tuning parameters for Finite Mixture of Accelerated Failure Time Regression Models
and Finite Mixture of Regression Models.
The penalties that are implemented in this package are \code{lasso}, \code{adplasso}, \code{scad}, \code{mcp}, \code{sica} and \code{hard}.
}
\examples{
\dontrun{Tuning Parameter Selection in Ovarian Cancer analysis

}
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
\references{
Shokoohi, F., Khalili, A., Asgharian, M. and Lin, S. (2016) Variable Selection in Mixture of Survival Models
}
\seealso{
Other lnorm..norm..weibull: \code{\link{fmrs.gen.data}},
  \code{\link{fmrs.mle}}, \code{\link{fmrs.varsel}}
}
\keyword{AFT,}
\keyword{Algorithm,}
\keyword{Censored}
\keyword{Data,}
\keyword{EM}
\keyword{FMR,}
\keyword{Regression}
\keyword{Ridge}

