% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMR_Surv_VarSel.R
\name{fmrs.varsel}
\alias{fmrs.varsel}
\title{Variable Selection in Finite Mixture of Accelerated Failure Time Regression Models
 and Finite Mixture of Regression Models}
\usage{
fmrs.varsel(y, x, delta, nComp, disFamily = "lnorm", initCoeff, initSigma,
  initPi, penFamily = "lasso", lambPen, lambRidge = 0, nIterEM = 2000,
  nIterNR = 2, conveps = 1e-08, convepsEM = 1e-08, convepsNR = 1e-08,
  porNR = 2, gamMixPor = 1)
}
\arguments{
\item{y}{Time-to-event observations}

\item{x}{Design matrix (covariates)}

\item{delta}{Censoring indicator vector}

\item{nComp}{Orde (Number of components) of mixture model}

\item{disFamily}{Specify sub-distributions family. The options are \code{"norm"} for FMR models,
\code{"lnorm"} for mixture of AFT regression models with Log-Normal sub-distributions,
\code{"weibull"} for mixture of AFT regression models with Weibull sub-distributions,}

\item{initCoeff}{Vector of initial values for coefficients including intercepts.}

\item{initSigma}{Vector of initial values for standard deviations}

\item{initPi}{Vector of initial values for proportion of components}

\item{penFamily}{The penalty that used in variable selection method.
The available options are  \code{"lasso"}, \code{"adplasso"}, \code{"mcp"}, \code{"scad"}, \code{"sica"} and \code{"hard"}.}

\item{lambPen}{A vector of lambda for penaly}

\item{lambRidge}{Lambda for ridge penalty}

\item{nIterEM}{Number of iteration for EM algorithm}

\item{nIterNR}{Number of iteration for Newton-Raphson algorithm}

\item{conveps}{A positive number for avoiding NaN in computing divisions.}

\item{convepsEM}{Treshold for convergence of EM algorithm}

\item{convepsNR}{Treshold for convergence of Newton-Raphson algorithm}

\item{porNR}{Used in pow(0.5, porNR) for tuning the increment in Newton-Raphson algorithm.}

\item{gamMixPor}{Proportion of mixing parameters in the penalty. The value must belong to the interval [0,1]. If \code{gamMixPor = 0}, the penalty structure is no longer mixture.}
}
\value{
Parameter estimates of mixture of Normal or mixture of Log-Normal AFT models

A fitted FMRs model object of class \code{\link{fmrs.fit}}
}
\description{
It provides variable selection and parameter estimation for Finite Mixture of Regression Models
and Finite Mixture of Accelerated Failure Time Regression Models with Log-Normal or Weibull sub-distributions.
The penalties lasso, scad, mcp, sica, adaptive lasso and hard are implemented. It also provide Ridge Regression and Elastic Net.
}
\examples{
\dontrun{Variable Selection in Ovarian Cancer analysis

}
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
\references{
Shokoohi, F., Khalili, A., Asgharian, M. and Lin, S. (2016) Variable Selection in Mixture of Survival Models
}
\seealso{
Other lnorm..norm..weibull: \code{\link{fmrs.gen.data}},
  \code{\link{fmrs.mle}}, \code{\link{fmrs.tunsel}}
}
\keyword{AFT,}
\keyword{Algorithm,}
\keyword{Censored}
\keyword{Data,}
\keyword{EM}
\keyword{FMR,}
\keyword{Regression}
\keyword{Ridge}

