% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{fmrstunpar-class}
\alias{fmrstunpar}
\alias{fmrstunpar-class}
\title{An S4 class to represent estimated optimal lambdas}
\description{
An S4 class to represent estimated optimal lambdas resulted
    from running \code{\link{fmrs.tunsel}}
}
\section{Slots}{

\describe{
\item{\code{ncomp}}{A length-one numeric vector}

\item{\code{lambPen}}{A dimension-one-\code{ncomp} numeric array}

\item{\code{disFamily}}{A length-one character vector}

\item{\code{penFamily}}{A length-one character vector}

\item{\code{lambRidge}}{A length-one numeric vector}

\item{\code{model}}{A length-one character vector}
}}
\keyword{object}

